Models 1.- Camio en la estructura de los modelos de la base de datos - Sustitución del valor Integer por BigInteger (estudiantes.py - materias.py - matriculas.py)

+++ db/models/estudiantes.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, Integer, String, BigInteger, Date
+from sqlalchemy import Column, String, BigInteger, Date
 from sqlalchemy.orm import relationship
 from ..conexion import Base
 
 class Estudiantes(Base):
     __tablename__ = "estudiantes"
 
-    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
     nombre = Column(String(20), nullable=False)
     apellido = Column(String(20), nullable=False)
     cedula = Column(BigInteger, nullable=False, unique=True)
